created helpers func
